requirements.txt
flask 
flask_cors 
supabase
moviepy
imageio-ffmpeg
requests
pytube
python-dotenv
groq